+++ backend/src/classes/Brave/Core/Migrations/Version20190318212603.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Brave\Core\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20190318212603 extends AbstractMigration
+{
+    public function getDescription() : string
+    {
+        return '';
+    }
+
+    /**
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    public function up(Schema $schema) : void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql("UPDATE players SET name = CONCAT(name, '#', id) WHERE name NOT LIKE '%#%'");
+    }
+
+    /**
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    public function down(Schema $schema) : void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+    }
+}