+++ frontend/src/pages/Home.vue
@@ -147,8 +147,8 @@
                 <div class="col-lg-4">
                     <div class="player-hdl">
                         <h2>Account</h2>
+                        <span class="text-muted">{{ player.name }}</span>
                         <span v-if="player.status === 'managed'" class="text-muted">(manually managed)</span>
-                        <span class="text-muted">Name: {{ player.name }}</span>
                     </div>
                     <div class="card border-secondary mb-3">
                         <h3 class="card-header">Groups</h3>